XWIKI-14289: Create an RSS feed for notifications * Move NotificationRSSRenderer#renderFeed to NotificationRSSManager#renderFeed * Update unit tests
