+++ xwiki-platform-core/xwiki-platform-notifications/xwiki-platform-notifications-api/src/main/resources/META-INF/components.txt
@@ -2,6 +2,7 @@ org.xwiki.notifications.internal.email.DefaultNotificationEmailRenderer
 org.xwiki.notifications.internal.email.NotificationEmailSender
 org.xwiki.notifications.internal.email.NotificationMimeMessageIterator
 org.xwiki.notifications.internal.email.NotificationUserIterator
+org.xwiki.notifications.internal.rss.DefaultNotificationRSSManager
 org.xwiki.notifications.internal.rss.DefaultNotificationRSSRenderer
 org.xwiki.notifications.internal.CachedModelBridge
 org.xwiki.notifications.internal.DefaultCompositeEventStatusManager
