+++ xwiki-platform-core/xwiki-platform-notifications/xwiki-platform-notifications-api/src/test/java/org/xwiki/notifications/internal/rss/DefaultNotificationRSSRendererTest.java
@@ -20,11 +20,8 @@
 package org.xwiki.notifications.internal.rss;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.Date;
 import java.util.HashSet;
-
-import javax.script.ScriptContext;
 
 import org.junit.Before;
 import org.junit.Rule;
@@ -32,21 +29,14 @@ import org.junit.Test;
 import org.xwiki.eventstream.Event;
 import org.xwiki.localization.ContextualLocalizationManager;
 import org.xwiki.model.reference.DocumentReference;
-import org.xwiki.model.reference.DocumentReferenceResolver;
 import org.xwiki.notifications.CompositeEvent;
-import org.xwiki.notifications.NotificationException;
-import org.xwiki.notifications.internal.ModelBridge;
-import org.xwiki.rendering.renderer.BlockRenderer;
-import org.xwiki.script.ScriptContextManager;
 import org.xwiki.template.Template;
 import org.xwiki.template.TemplateManager;
 import org.xwiki.test.mockito.MockitoComponentMockingRule;
 
 import com.rometools.rome.feed.synd.SyndEntry;
-import com.rometools.rome.feed.synd.SyndFeed;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -65,32 +55,14 @@ public class DefaultNotificationRSSRendererTest
 
     private ContextualLocalizationManager contextualLocalizationManager;
 
-    private ModelBridge modelBridge;
-
-    private DocumentReferenceResolver<String> documentReferenceResolver;
-
     private TemplateManager templateManager;
-
-    private ScriptContextManager scriptContextManager;
-
-    private BlockRenderer blockRenderer;
 
     @Before
     public void setUp() throws Exception
     {
         this.contextualLocalizationManager = this.mocker.registerMockComponent(ContextualLocalizationManager.class);
 
-        this.modelBridge = this.mocker.registerMockComponent(ModelBridge.class);
-
-        this.documentReferenceResolver = this.mocker.registerMockComponent(DocumentReferenceResolver.class);
-
         this.templateManager = this.mocker.registerMockComponent(TemplateManager.class);
-
-        this.scriptContextManager = this.mocker.registerMockComponent(ScriptContextManager.class);
-        ScriptContext scriptContext = mock(ScriptContext.class);
-        when(this.scriptContextManager.getCurrentScriptContext()).thenReturn(scriptContext);
-
-        this.blockRenderer = this.mocker.registerMockComponent(BlockRenderer.class, "html/5.0");
     }
 
     private void mockEvent(CompositeEvent testCompositeEvent) throws Exception
@@ -126,23 +98,5 @@ public class DefaultNotificationRSSRendererTest
         assertEquals("TranslatedEventTitle", resultEntry.getTitle());
         assertEquals(1, resultEntry.getAuthors().size());
         assertEquals("id1", resultEntry.getUri());
-    }
-
-    @Test
-    public void renderFeed() throws Exception
-    {
-        when(this.contextualLocalizationManager.getTranslationPlain("notifications.rss.feedTitle"))
-                .thenReturn("FeedTitle");
-        when(this.modelBridge.getDocumentURL(
-                any(), any(), any())).thenReturn("url");
-        when(this.contextualLocalizationManager.getTranslationPlain("notifications.rss.feedDescription"))
-                .thenReturn("FeedDescription");
-
-        SyndFeed feed = this.mocker.getComponentUnderTest().renderFeed(Collections.EMPTY_LIST);
-
-        assertEquals("FeedTitle", feed.getTitle());
-        assertEquals("FeedDescription", feed.getDescription());
-        assertEquals("url", feed.getLink());
-        assertEquals(0, feed.getEntries().size());
     }
 }
